rearrange: create add module
